requirements.txt
aiohttp==3.6.2
requests==2.22.0
redis==3.4.0
aiofiles==0.5.0
motor==2.1.0
python_dateutil==2.8.1
